Absort ssplit, get compile working
