+++ CMakeLists.txt
@@ -20,12 +20,14 @@ else(USE_SYSTEM_LIBS)
   add_subdirectory(3rd-party)
 endif(USE_SYSTEM_LIBS)
 
+find_package(PCRE2 REQUIRED)
+
 set(SLIMT_COMPILE_OPTIONS -Wall -Wpedantic -Wextra -Wno-unknown-pragmas -Werror)
 set(SLIMT_COMPILE_OPTIONS ${SLIMT_COMPILE_OPTIONS}
                           -Wno-missing-field-initializers -Wno-unused-function)
 set(SLIMT_COMPILE_DEFINITIONS "")
 
-set(SLIMT_EXTERNAL_LIBS SentencePiece::SentencePiece)
+set(SLIMT_EXTERNAL_LIBS SentencePiece::SentencePiece ${PCRE2_LIBRARIES})
 
 if(WITH_BLAS)
   find_package(BLAS)
