+++ slimt/Splitter.hh
@@ -26,7 +26,7 @@ class Splitter {
   explicit Splitter(const std::string& prefix_file);
 
   void load(const std::string& fname);
-  void loadFromSerialized(const std::string_view buffer);
+  void load_from_serialized(const std::string_view buffer);
 
   // Find next sentence boundary, return StringPiece for next sentence,
   // advance rest to reflect the rest of the text.
